requirements.txt
# DESS - Desarrollador de Ecosistemas de Soluciones Empresariales
# Requirements del Estado Actual del Sistema - Julio 2025

# ============================================================================
# DEPENDENCIAS CORE - FRAMEWORK PRINCIPAL
# ============================================================================

# Framework Django
Django==4.2
python-decouple==3.8

# Base de Datos (Actualmente usando SQLite, Oracle disponible)
oracledb==1.4.0

# ============================================================================
# DEPENDENCIAS API Y DOCUMENTACIÓN
# ============================================================================

# API Documentation (OpenAPI/Swagger)
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5

# ============================================================================
# DEPENDENCIAS DE AUTENTICACIÓN Y SEGURIDAD
# ============================================================================

# JWT para autenticación API
PyJWT==2.9.0

# Cliente HTTP para APIs externas
requests==2.32.4

# Excel export functionality
openpyxl==3.1.2

# ============================================================================
# DEPENDENCIAS CORE DJANGO
# ============================================================================

# ASGI para aplicaciones asíncronas
asgiref==3.8.1

# Parser SQL para Django
sqlparse==0.5.3

# Zona horaria para Django
tzdata==2025.2

# Type hints para Python < 3.10
typing_extensions==4.14.0

# ============================================================================
# DEPENDENCIAS DE UTILIDADES
# ============================================================================

# Parsing YAML para configuraciones
PyYAML==6.0.2

# Inflections para nombres de API
inflection==0.5.1

# Generación de templates URI
uritemplate==4.2.0

# Validación de referencias JSON Schema
referencing==0.36.2
rpds-py==0.25.1

# ============================================================================
# DEPENDENCIAS DE RED Y CERTIFICADOS
# ============================================================================

# Certificados SSL/TLS
certifi==2025.6.15

# Codificación de caracteres
charset-normalizer==3.4.2

# URLs internacionales
idna==3.10

# Cliente HTTP seguro
urllib3==2.5.0

# ============================================================================
# DEPENDENCIAS DE TESTING Y VALIDACIÓN
# ============================================================================

# Framework de testing y validación
attrs==25.3.0

# ============================================================================
# DEPENDENCIAS OPCIONALES - DESARROLLO
# ============================================================================

# Las siguientes dependencias están comentadas ya que no son estrictamente
# necesarias para el funcionamiento core del sistema pero pueden ser útiles
# para desarrollo y testing:

# Testing Framework
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code Quality Tools
black==23.11.0
flake8==6.1.0

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Dependency Injection Framework
dependency-injector==4.41.0

# Git Integration para sistema de deployment
GitPython==3.1.40

# Process Management
psutil==5.9.6

# ============================================================================
# DEPENDENCIAS DE DESPLIEGUE Y CONTENEDORES
# ============================================================================

# Docker SDK para gestión de contenedores
docker==7.0.0

# CORS para APIs de webhook
django-cors-headers==4.3.1

# Servidor WSGI para producción
gunicorn==21.2.0

# Servir archivos estáticos en producción
whitenoise==6.6.0

# Cache con Redis
django-redis==5.4.0
redis==5.0.1

# Oracle Database Driver (para ambiente de producción)

# ============================================================================
# NOTAS IMPORTANTES
# ============================================================================
#
# 1. Este archivo refleja el estado ACTUAL del proyecto DESS
# 2. Dependencias marcadas como "comentadas" están disponibles en el documento
#    de requerimientos de modernización si se necesitan en el futuro
# 3. El sistema actualmente usa SQLite para desarrollo, Oracle está configurado
#    pero no activo
# 4. Todas las dependencias listadas están validadas como necesarias para
#    el funcionamiento actual del sistema
# 5. Para ambiente de producción, descomentar cx-Oracle y configurar DB
# 1. Este archivo refleja el estado ACTUAL del proyecto DESS
# 2. Dependencias marcadas como "comentadas" están disponibles en el documento
#    de requerimientos de modernización si se necesitan en el futuro
# 3. El sistema actualmente usa SQLite para desarrollo, Oracle está configurado
#    pero no activo
# 4. Todas las dependencias listadas están validadas como necesarias para
#    el funcionamiento actual del sistema
# 5. Para ambiente de producción, descomentar cx-Oracle y configurar DB
